requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
dj-database-url==2.3.0
Django==5.2
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.10
inflection==0.5.1
packaging==25.0
Pillow==10.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.2
requests==2.31.0
setuptools==78.1.1
sib-api-v3-sdk==7.6.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
whitenoise==6.6.0
